move LifecycleOwner & SavedStateRegistryOwner codes to commonMain
